+++ libs/designer/src/lib/core/state/operation/operationSelector.ts
@@ -29,6 +29,24 @@ export const useParameterStaticResult = (nodeId: string): NodeStaticResults => {
     return rootState.operations.staticResults[nodeId];
   });
 };
+export const useTokenDependencies = (nodeId: string) =>
+  useSelector((rootState: RootState) => {
+    const operationInputParameters = rootState.operations.inputParameters[nodeId];
+    if (!operationInputParameters) {
+      return new Set();
+    }
+    const dependencies = new Set();
+    for (const group of Object.values(operationInputParameters.parameterGroups)) {
+      for (const parameter of group.parameters) {
+        for (const value of parameter.value) {
+          if (value.token?.actionName) {
+            dependencies.add(value.token.actionName);
+          }
+        }
+      }
+    }
+    return dependencies;
+  });
 
 export const useParameterValidationErrors = (nodeId: string) => {
   return useSelector((rootState: RootState) => {